requirements.txt
moviepy==2.1.1
PyPDF2==3.0.1
google-generativeai==0.8.3
gTTS==2.5.4
pydub==0.25.1
python-dotenv==1.0.1